A parser to read total ion chromatogram data streams from 'Shimadzu' .lcd files. LCD files are encoded as 'Microsoft' OLE documents. The parser relies on the olefile package in Python to unpack the files. The TIC data is encoded in a stream called Centroid:3D Raw Data. The PDA data stream contains a segment for each retention time, beginning with a 8-byte header. After the header, the file consists of a series of 4-byte little-endian integers in blocks of 3 (16-bytes per block), followed by a 4-byte spacer (00000000) The first integer is the retention time (scaled by 1000), the second integer is the scan number, and the third integer is the intensity.

A parser to read total ion chromatogram data streams from 'Shimadzu' .lcd files. LCD files are encoded as 'Microsoft' OLE documents. The parser relies on the olefile package in Python to unpack the files. The TIC data is encoded in a stream called Centroid:3D Raw Data. The PDA data stream contains a segment for each retention time, beginning with a 8-byte header. After the header, the file consists of a series of 4-byte little-endian integers in blocks of 3 (16-bytes per block), followed by a 4-byte spacer (00000000) The first integer is the retention time (scaled by 1000), the second integer is the scan number, and the third integer is the intensity.

Usage

read_sz_tic(
  path,
  format_out = c("matrix", "data.frame"),
  data_format = c("wide", "long"),
  read_metadata = TRUE
)

Arguments

path

Path to LCD file.

format_out

Matrix or data.frame.

data_format

Either wide (default) or long.

read_metadata

Logical. Whether to attach metadata.

Value

A 2D chromatogram from the SumTIC stream in matrix or data.frame format, according to the value of format_out. The chromatograms will be returned in wide or long format according to the value of data_format.

Note

This parser is experimental and may still need some work. It is not yet able to interpret much metadata from the files.

Author

Ethan Bass