Writes OpenChrom XML batch file This function is called internally by call_openchrom.

Writes OpenChrom XML batch file This function is called internally by call_openchrom.

Usage

write_openchrom_batchfile(
  files,
  path_out,
  format_in = c("msd", "csd", "wsd"),
  export_format = c("csv", "cdf", "mzml", "animl")
)

Arguments

files

Paths to files for conversion

path_out

directory to export converted files.

format_in

Either msd for mass spectrometry data, csd for flame ionization data, or wsd for DAD/UV data.

export_format

Either csv, cdf, mzml, or animl.

Value

Returns path to newly created xml batch file.

Author

Ethan Bass